TPC 09:
Changed class AbstractBindMember: Modified signature of Method AbstractBindMember.addFormatter as to include TargetClass as a paramemter. This will allow us to get an instance of the method of the target class we wish to use as a formatter method and create a new, anonymous class of Type Formatter wich will invoke this method upon invocation of its "Formatter.format" method.

This way, we can have a single Map<AnnotatedElement, Formatter> as a means of cacheing these values and thus dramatically reducing the need for use of reflexion.

Also, changed classes BindField and BindProp (both constructors) as to match the new AbstractBinder.addFormatter signature.